除了垃圾清理，功能大致完成，待修复bug

+++ app/src/main/java/com/lzokks04/myutilbox/adapter/AppInfoAdapter.java
@@ -42,7 +42,7 @@ public class AppInfoAdapter extends RecyclerView.Adapter<AppInfoAdapter.MyViewHo
         holder.ivIcon.setImageDrawable(appInfoList.get(position).getIcon());
         holder.tvAppname.setText(appInfoList.get(position).getName());
         holder.tvPackname.setText(appInfoList.get(position).getPackageName());
-        holder.tvSize.setText("大小:" + Formatter.formatFileSize(context, (long) appInfoList.get(position).getSize()));
+        holder.tvSize.setText("大小:" + Formatter.formatFileSize(context, appInfoList.get(position).getSize()));
     }
 
     @Override
